server/requirements.txt
airtable-python-wrapper==0.15.3
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
